requirements-docs.txt
# Optional: regenerate PNG screenshots for the README (faculty / reports).
pillow>=10.0.0
